enablement/assignment_3/requirements.txt
sentence-transformers
chromadb
openai
python-dotenv